// src/app/enum/key.enum.ts
export enum Key {
    Z = 'KeyZ',
    X = 'KeyX',
    SPACE = 'Space',
    LEFT = 'ArrowLeft',
    RIGHT = 'ArrowRight',
    UP = 'ArrowUp',
    DOWN = 'ArrowDown',
    Q = 'KeyQ',
    W = 'KeyW',
    E = 'KeyE',
    A = 'KeyA',
    S = 'KeyS',
    D = 'KeyD',
    ESCAPE = 'Escape',
    ENTER = 'Enter'
}